IT RUNS! All 7 loops, all stimuli!
Fix was related to the length of the strings in the SIT condition: the longest sentence was 85 characters, which was longer than the longest string variable in the data file ("monologue"). When the script tried to concatenate the data file and the new stimuli file, it would fail if the new stimuli contained a string variable longer than what was on the full data file (the matrices were of unequal dimensions). To fix this, I edited the "monologue" text so it was longer than any of the SIT sentences (thus making the default length as long as it could be).  A bit hacky, and probably could be avoided by somehow recoding the variables, but when I tried to make them char or nominal arrays the script would fail, so this is just a hacky way of getting it to do what I want, mostly.